# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SessionGuard API",
    "version": "1.5.2"
  },
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Session Events",
        "description": "Return all events for a session \u2014 full timeline for charts.\nAdds spin_number and cumulative balance column.",
        "operationId": "get_session_events_api_v1_events_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/summary": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Events Summary",
        "description": "Aggregated event statistics \u2014 used by session detail KPIs.",
        "operationId": "get_events_summary_api_v1_events_summary_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/validate/{session_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Validate Events",
        "description": "Validate all events for a session and return flagged issues.",
        "operationId": "validate_events_api_v1_events_validate__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/coach/coach/{run_id}": {
      "get": {
        "tags": [
          "coach"
        ],
        "summary": "Get Coach Message",
        "operationId": "get_coach_message_api_v1_coach_coach__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "balanced",
              "title": "Style"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/coach/coach/{run_id}/reset": {
      "post": {
        "tags": [
          "coach"
        ],
        "summary": "Reset Coach State",
        "operationId": "reset_coach_state_api_v1_coach_coach__run_id__reset_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/import-wizard/import/preview": {
      "post": {
        "tags": [
          "import-wizard"
        ],
        "summary": "Preview Import",
        "description": "Upload a CSV, store it, return columns + preview rows + suggested mapping.",
        "operationId": "preview_import_api_v1_import_wizard_import_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import_api_v1_import_wizard_import_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Preview Import Api V1 Import Wizard Import Preview Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/import-wizard/import/confirm": {
      "post": {
        "tags": [
          "import-wizard"
        ],
        "summary": "Confirm Import",
        "description": "Apply the column mapping, create a session, insert events, return counts.",
        "operationId": "confirm_import_api_v1_import_wizard_import_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Confirm Import Api V1 Import Wizard Import Confirm Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_preview_import_api_v1_import_wizard_import_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_import_api_v1_import_wizard_import_preview_post"
      },
...
      "ColumnMapping": {
        "properties": {
          "date": {
            "type": "string",
            "minLength": 1,
            "title": "Date"
          },
          "bet": {
            "type": "string",
            "minLength": 1,
            "title": "Bet"
          },
          "win": {
            "type": "string",
            "minLength": 1,
            "title": "Win"
          },
          "balance": {
            "type": "string",
            "minLength": 1,
            "title": "Balance"
          },
          "spin_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spin Number"
          }
        },
        "type": "object",
        "required": [
          "date",
          "bet",
          "win",
          "balance"
        ],
        "title": "ColumnMapping"
      },
...
      "ConfirmImportRequest": {
        "properties": {
          "upload_id": {
            "type": "string",
            "minLength": 1,
            "title": "Upload Id"
          },
          "column_mapping": {
            "$ref": "#/components/schemas/ColumnMapping"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "column_mapping"
        ],
        "title": "ConfirmImportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}